fix Level

+++ Signum.Engine.Extensions/Tree/TreeLogic.cs
@@ -256,7 +256,7 @@ namespace Signum.Engine.Tree
                 c.Construct = (t, _) => new T
                 {
                     ParentOrSibling = t.ToLite(),
-                    Level = t.Level + 1,
+                    Level = (short)(t.Level + 1),
                     IsSibling = false
                     //                    
                 };
